# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cloud-api/onboarding/webhooks/{provider}": {
      "get": {
        "description": "Handles webhook subscription verification from Meta (WhatsApp) and Twilio (SMS).\n\n    For WhatsApp: Returns hub.challenge if hub.verify_token matches.\n    For SMS: Returns challenge if verify_token matches.",
        "operationId": "VerificationWebhookController_verifyWebhook",
        "parameters": [
          {
            "name": "hub.mode",
            "required": false,
            "in": "query",
            "description": "WhatsApp: Subscription mode (should be \"subscribe\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.challenge",
            "required": false,
            "in": "query",
            "description": "WhatsApp: Challenge string to echo back",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.verify_token",
            "required": false,
            "in": "query",
            "description": "WhatsApp: Verification token to validate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "verify_token",
            "required": false,
            "in": "query",
            "description": "SMS/Twilio: Verification token to validate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "challenge",
            "required": false,
            "in": "query",
            "description": "SMS/Twilio: Challenge string to echo back",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "Webhook provider type",
            "schema": {
              "enum": [
                "whatsapp",
                "sms"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook verified successfully, returns the challenge string",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported provider"
          },
          "401": {
            "description": "Invalid verification token"
          }
        },
        "summary": "Verify webhook endpoint (Meta/Twilio subscription verification)",
        "tags": [
          "Onboarding - Webhooks"
        ]
      },
      "post": {
        "description": "Receives and processes incoming message webhooks from Meta (WhatsApp) and Twilio (SMS).\n\n    Validates webhook signatures before processing:\n    - WhatsApp: Uses X-Hub-Signature-256 header (HMAC-SHA256)\n    - SMS: Uses X-Twilio-Signature header (HMAC-SHA1)\n\n    Messages containing verification tokens are processed to complete mobile verification.",
        "operationId": "VerificationWebhookController_handleWebhook",
        "parameters": [
          {
            "name": "x-hub-signature-256",
            "required": false,
            "in": "header",
            "description": "WhatsApp webhook signature (HMAC-SHA256)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-twilio-signature",
            "required": false,
            "in": "header",
            "description": "Twilio SMS webhook signature (HMAC-SHA1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "Webhook provider type",
            "schema": {
              "enum": [
                "whatsapp",
                "sms"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Webhook payload from provider (WhatsApp or Twilio format)",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "description": "WhatsApp webhook payload",
                    "properties": {
                      "object": {
                        "type": "string",
                        "example": "whatsapp_business_account"
                      },
                      "entry": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "changes": {
                              "type": "array"
                            }
                          }
                        }
                      }
                    }
                  },
                  {
                    "type": "object",
                    "description": "Twilio SMS webhook payload",
                    "properties": {
                      "MessageSid": {
                        "type": "string"
                      },
                      "From": {
                        "type": "string"
                      },
                      "To": {
                        "type": "string"
                      },
                      "Body": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook received and queued for processing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unsupported provider or malformed payload"
          },
          "401": {
            "description": "Invalid webhook signature"
          }
        },
        "summary": "Handle incoming webhook events from WhatsApp or SMS providers",
        "tags": [
          "Onboarding - Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti Cloud API",
    "description": "Internal API for Vritti SaaS Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
...
  }
}